requirements.txt
aiodns==2.0.0
aiohttp==3.7.4.post0
alabaster==0.7.12
async-timeout==3.0.1
attrs @ file:///tmp/build/80754af9/attrs_1604765588209/work
Babel @ file:///tmp/build/80754af9/babel_1607110387436/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
Brotli @ file:///opt/concourse/worker/volumes/live/6f34e4d0-4572-4e67-52c6-900349b58bc0/volume/brotli-split_1602517645446/work
brotlipy==0.7.0
ccxt==1.48.30
certifi==2020.12.5
cffi @ file:///opt/concourse/worker/volumes/live/ca3c0fea-9d0f-4489-6889-89b42078ed37/volume/cffi_1613246931142/work
chardet @ file:///opt/concourse/worker/volumes/live/9efbf151-b45b-463d-6340-a5c399bf00b7/volume/chardet_1607706825988/work
click @ file:///home/linux1/recipes/ci/click_1610990599742/work
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
cryptography @ file:///opt/concourse/worker/volumes/live/cdf8ff17-b0bf-4081-524c-b5a0afe929ba/volume/cryptography_1616769280208/work
cycler==0.10.0
dash @ file:///tmp/build/80754af9/dash_1611168717544/work
dash-core-components==1.3.1
dash-html-components==1.0.1
dash-renderer==1.1.2
dash-table==4.4.1
decorator @ file:///home/ktietz/src/ci/decorator_1611930055503/work
docutils==0.16
et-xmlfile==1.0.1
Flask @ file:///home/ktietz/src/ci/flask_1611932660458/work
Flask-Compress @ file:///tmp/build/80754af9/flask-compress_1613587450386/work
future==0.18.2
gevent @ file:///opt/concourse/worker/volumes/live/c9692c6d-50c4-4197-4de2-6c138d206ad6/volume/gevent_1616772940935/work
greenlet @ file:///opt/concourse/worker/volumes/live/e96eb331-0f3d-4286-759a-ce903db4e49e/volume/greenlet_1611958360647/work
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
imagesize @ file:///home/ktietz/src/ci/imagesize_1611921604382/work
importlib-metadata @ file:///opt/concourse/worker/volumes/live/4e25a0be-45cc-4b73-4314-4604f00e30a4/volume/importlib-metadata_1617877365451/work
iniconfig @ file:///home/linux1/recipes/ci/iniconfig_1610983019677/work
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2 @ file:///tmp/build/80754af9/jinja2_1612213139570/work
kiwisolver @ file:///opt/concourse/worker/volumes/live/1e6d28e8-da14-4365-7344-567a05db7efc/volume/kiwisolver_1612282399491/work
lxml @ file:///opt/concourse/worker/volumes/live/a4879036-1466-48b6-6c8d-95f5cf7bc17d/volume/lxml_1616443217052/work
MarkupSafe==1.1.1
matplotlib @ file:///opt/concourse/worker/volumes/live/41e8cd50-031f-4dda-5787-dd3c4f4e0f08/volume/matplotlib-suite_1613407855571/work
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
more-itertools @ file:///tmp/build/80754af9/more-itertools_1613676688952/work
multidict==5.1.0
numpy @ file:///opt/concourse/worker/volumes/live/a1b5ea96-b05c-40ba-5dfb-4542ece1c186/volume/numpy_and_numpy_base_1603491228242/work
olefile==0.46
openpyxl @ file:///tmp/build/80754af9/openpyxl_1615411699337/work
packaging @ file:///tmp/build/80754af9/packaging_1611952188834/work
pandas==1.2.3
pandas-datareader @ file:///tmp/build/80754af9/pandas-datareader_1595379313840/work
pep8==1.7.1
Pillow @ file:///opt/concourse/worker/volumes/live/a50650a0-aa5f-49d7-67b4-33442af0a044/volume/pillow_1615057387162/work
plotly @ file:///tmp/build/80754af9/plotly_1610565580232/work
pluggy @ file:///opt/concourse/worker/volumes/live/f85b5d5a-abad-4f46-46ad-4ead594efa77/volume/pluggy_1615976299968/work
ptyprocess==0.6.0
py @ file:///tmp/build/80754af9/py_1607971587848/work
pycares==3.1.1
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments @ file:///tmp/build/80754af9/pygments_1615143339740/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
PySocks @ file:///opt/concourse/worker/volumes/live/ef943889-94fc-4539-798d-461c60b77804/volume/pysocks_1605305801690/work
pytest==6.2.2
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
PyYAML==5.4.1
pyzmq==19.0.2
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
retrying==1.3.3
scipy @ file:///opt/concourse/worker/volumes/live/cd12a34b-67bb-4e81-5e11-d0d209d4b37d/volume/scipy_1616703203371/work
seaborn @ file:///tmp/build/80754af9/seaborn_1608578541026/work
six @ file:///opt/concourse/worker/volumes/live/f983ba11-c9fe-4dff-7ce7-d89b95b09771/volume/six_1605205318156/work
snowballstemmer @ file:///tmp/build/80754af9/snowballstemmer_1611258885636/work
Sphinx @ file:///tmp/build/80754af9/sphinx_1616268783226/work
sphinxcontrib-applehelp @ file:///home/ktietz/src/ci/sphinxcontrib-applehelp_1611920841464/work
sphinxcontrib-devhelp @ file:///home/ktietz/src/ci/sphinxcontrib-devhelp_1611920923094/work
sphinxcontrib-htmlhelp @ file:///home/ktietz/src/ci/sphinxcontrib-htmlhelp_1611920974801/work
sphinxcontrib-jsmath @ file:///home/ktietz/src/ci/sphinxcontrib-jsmath_1611920942228/work
sphinxcontrib-qthelp @ file:///home/ktietz/src/ci/sphinxcontrib-qthelp_1611921055322/work
sphinxcontrib-serializinghtml @ file:///home/ktietz/src/ci/sphinxcontrib-serializinghtml_1611920755253/work
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
tornado==6.0.4
typing-extensions @ file:///home/ktietz/src/ci_mi/typing_extensions_1612808209620/work
urllib3 @ file:///tmp/build/80754af9/urllib3_1615837158687/work
websocket-client @ file:///opt/concourse/worker/volumes/live/27763808-0b51-45c4-7a46-77805b7949ef/volume/websocket-client_1614804248608/work
Werkzeug @ file:///home/ktietz/src/ci/werkzeug_1611932622770/work
xlrd @ file:///tmp/build/80754af9/xlrd_1608072521494/work
yarl==1.1.0
zipp @ file:///tmp/build/80754af9/zipp_1615904174917/work
zope.event==4.5.0
zope.interface @ file:///opt/concourse/worker/volumes/live/c60b05dd-b3e4-49e9-4328-f84992cb3b8c/volume/zope.interface_1616357196088/work
